+++ R/fuse.R
@@ -465,7 +465,11 @@ fuse = function(input, output = NULL, text = NULL, envir = parent.frame(), quiet
   oenv = as.list(.env); on.exit(reset_env(oenv, .env), add = TRUE)
   # use default fig.path/cache.path instead of inheriting (#127)
   nested = isTRUE(.env$in_fuse); .env$in_fuse = TRUE
-  if (nested) opts$fig.path = opts$cache.path = NULL
+  if (nested) {
+    old_paths = list(fig.path = opts$fig.path, cache.path = opts$cache.path)
+    opts$fig.path = opts$cache.path = NULL
+    on.exit(reactor(old_paths), add = TRUE)
+  }
 
   # set working directory if unset
   if (is_file(input) && is.null(opts$wd)) opts$wd = dirname(normalizePath(input))